+++ loader/src/platform/mac/input.mm
@@ -284,7 +284,11 @@ void keyDownExecHook(EAGLView* self, SEL sel, NSEvent* event) {
         modifiers & KeyboardModifier::Super
     );
 
-    NSString* characters = [event charactersIgnoringModifiers];
+    NSString* characters = [event characters];
+    if ([characters length] == 0) {
+        characters = [event charactersIgnoringModifiers];
+    }
+
     if ([characters length] != 0) {
         unichar character = [characters characterAtIndex:0];
         if (character == NSDeleteFunctionKey || character == 0x7f) {
@@ -423,14 +427,19 @@ void rightMouseDownHook(EAGLView* self, SEL sel, NSEvent* event) {
         modifiers
     );
 
-    if (MouseInputEvent().send(data) != ListenerResult::Propagate) return;
+    Loader::get()->queueInMainThread([data = std::move(data)] mutable {
+        MouseInputEvent().send(data);
+    });
 
     [self performSelector:sel withObject:event];
 }
 
 void rightMouseDraggedHook(EAGLView* self, SEL sel, NSEvent* event) {
     NSPoint point = [self convertPoint:[event locationInWindow] fromView:nil];
-    if (MouseMoveEvent().send(point.x, point.y) != ListenerResult::Propagate) return;
+
+    Loader::get()->queueInMainThread([point] {
+        MouseMoveEvent().send(point.x, point.y);
+    });
 
     [self performSelector:sel withObject:event];
 }
@@ -445,7 +454,9 @@ void rightMouseUpHook(EAGLView* self, SEL sel, NSEvent* event) {
         modifiers
     );
 
-    if (MouseInputEvent().send(data) != ListenerResult::Propagate) return;
+    Loader::get()->queueInMainThread([data = std::move(data)] mutable {
+        MouseInputEvent().send(data);
+    });
 
     [self performSelector:sel withObject:event];
 }
@@ -460,14 +471,19 @@ void otherMouseDownHook(EAGLView* self, SEL sel, NSEvent* event) {
         modifiers
     );
 
-    if (MouseInputEvent().send(data) != ListenerResult::Propagate) return;
+    Loader::get()->queueInMainThread([data = std::move(data)] mutable {
+        MouseInputEvent().send(data);
+    });
 
     [self performSelector:sel withObject:event];
 }
 
 void otherMouseDraggedHook(EAGLView* self, SEL sel, NSEvent* event) {
     NSPoint point = [self convertPoint:[event locationInWindow] fromView:nil];
-    if (MouseMoveEvent().send(point.x, point.y) != ListenerResult::Propagate) return;
+
+    Loader::get()->queueInMainThread([point] {
+        MouseMoveEvent().send(point.x, point.y);
+    });
 
     [self performSelector:sel withObject:event];
 }
@@ -482,7 +498,9 @@ void otherMouseUpHook(EAGLView* self, SEL sel, NSEvent* event) {
         modifiers
     );
 
-    if (MouseInputEvent().send(data) != ListenerResult::Propagate) return;
+    Loader::get()->queueInMainThread([data = std::move(data)] mutable {
+        MouseInputEvent().send(data);
+    });
 
     [self performSelector:sel withObject:event];
 }